+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Changelog for NeoFS Node
 - SN does not retry resending failed transaction because of insufficient GAS in some cases (#3839)
 - Too early GET/HEAD/RANGE request failure on single SN dial failure (#3840)
 - Payment deadlock on IR side (#3842)
-- Wrong shard ID entry in FSTree descriptor (#3849)
+- Wrong shard ID entry in FSTree descriptor (#3849, #3861)
 
 ### Changed
 - SN returns unsigned responses to requests with API >= `v2.22` (#3785)